increase fraud detection 2 cell timeout (#52)
We were getting timeouts on the `fraud-detection-2` notebook in CI that
had a 30m cell timeout. On Paperspace, the cell ran in 29m 54s, so there
was not much buffer, and so on CI we also see that it sometimes passes
and sometimes fails (I'm assuming due to system-level behaviour that
affects e.g. dataloading speed). This PR increases the timeout to
provide more buffer, verified in CI that the test now passes:
https://github.com/graphcore/Gradient-Pytorch-Geometric-private/actions/runs/5854317396